Started working on stats

+++ IdleRpg/Game/StatModifier.cs
@@ -6,5 +6,5 @@ public class StatModifier
 {
     public required Enum Stat { get; init; }
     public List<Enum> StatsUsed { get; init; } = new();
-    public required Func<Dictionary<Enum, int>, int> Calculation { get; init; }
+    public required Func<Dictionary<Enum, long>, long> Calculation { get; init; }
 }
